Initial Commit
Initial Commit

Note: I WON'T BE UPDATING THIS. THIS IS FOR THE COMMUNITY TO UPDATE.

It is not my responsibility to compensate for older badly coded systems.
But you all can help each other by updating this. If I see people
working with this, I will help when I can.